chore(tauri) move endpoint specific modules
