+++ tauri/src/salt.rs
@@ -2,7 +2,6 @@ use std::sync::Mutex;
 
 use lazy_static::lazy_static;
 use uuid::Uuid;
-use web_view::WebView;
 
 struct Salt {
   value: String,
@@ -48,21 +47,4 @@ pub fn is_valid(salt: String) -> bool {
     }
     None => false,
   }
-}
-
-pub fn validate<T: 'static>(
-  webview: &mut WebView<'_, T>,
-  salt: String,
-  callback: String,
-  error: String,
-) {
-  let response = if is_valid(salt) {
-    Ok("'VALID'".to_string())
-  } else {
-    Err("'INVALID SALT'".to_string())
-  };
-  let callback_string = crate::api::rpc::format_callback_result(response, callback, error);
-  webview
-    .eval(callback_string.as_str())
-    .expect("Failed to eval JS from validate()");
 }
